use query inputs for model values
